// app/generated/prisma/models/tb_users.ts
export type tb_usersUpdateInput = {
  full_name?: Prisma.StringFieldUpdateOperationsInput | string
  username?: Prisma.StringFieldUpdateOperationsInput | string
  phone?: Prisma.StringFieldUpdateOperationsInput | string
  email?: Prisma.StringFieldUpdateOperationsInput | string
  password?: Prisma.StringFieldUpdateOperationsInput | string
  avatar_url?: Prisma.NullableStringFieldUpdateOperationsInput | string | null
  is_deleted?: Prisma.NullableBoolFieldUpdateOperationsInput | boolean | null
  created_at?: Prisma.NullableDateTimeFieldUpdateOperationsInput | Date | string | null
  updated_at?: Prisma.NullableDateTimeFieldUpdateOperationsInput | Date | string | null
  confirm_rules?: Prisma.NullableBoolFieldUpdateOperationsInput | boolean | null
  confirm_phone?: Prisma.NullableBoolFieldUpdateOperationsInput | boolean | null
  confirm_email?: Prisma.NullableBoolFieldUpdateOperationsInput | boolean | null
  tb_conversations?: Prisma.tb_conversationsUpdateManyWithoutTb_usersNestedInput
  tb_projects?: Prisma.tb_projectsUpdateManyWithoutTb_usersNestedInput
  tb_transactions?: Prisma.tb_transactionsUpdateManyWithoutTb_usersNestedInput
}